% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chromosomes.R
\name{add_chromosomes}
\alias{add_chromosomes}
\title{Add chromosomes to plot}
\usage{
add_chromosomes(
  p,
  y = -0.05,
  offset = 0.01,
  size = 2,
  color = "black",
  unit = "location"
)
}
\arguments{
\item{p}{A ggplot showing a genome scan (see \code{?plot_genome_scan}).}

\item{y}{Vertical location of the chromosomes on the plot.}

\item{offset}{Small number controlling the distance between the chromosomes.}

\item{size}{Width of the bars.}

\item{color}{Color of the bars.}

\item{unit}{Whether the horizontal axis is "locus" or "location" (see \code{?plot_genome_scan}).}
}
\value{
A ggplot
}
\description{
Add chromosomes as lines to a genome scan.
}
