% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geom_line_colored.R
\name{add_geom_line_colored}
\alias{add_geom_line_colored}
\title{Customize line colors}
\usage{
add_geom_line_colored(
  p,
  color_by = NULL,
  color_by_numeric = FALSE,
  colors = NULL
)
}
\arguments{
\item{p}{A ggplot}

\item{color_by}{The variable to use for coloration}

\item{color_by_numeric}{Whether to convert the color variable into a numeric}

\item{colors}{Optional colors of the lines. If color_by is NULL, the color that all the lines must have. If color_by is defined and color_by_numeric is TRUE, the lower and the upper end of the color gradient. If color_by_numeric is FALSE, a set of colors for each level of the color factor. If not defined, default ggplot color(s) will be used.}

\item{return}{A ggplot with colored lines added}
}
\description{
Easy way to add lines to a ggplot and color them with custom colors and variables
}
