% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_parameters.R
\name{collect_parameters}
\alias{collect_parameters}
\title{Collect parameters from multiple simulations}
\usage{
collect_parameters(
  simulations,
  parnames = NULL,
  pb = TRUE,
  filename = "paramlog.txt"
)
}
\arguments{
\item{simulations}{Paths to the simulations}

\item{parnames}{optional vector of parameter names (collects all parameters if not specified)}

\item{pb}{Whether to display a progress bar}

\item{filename}{Optional file name}
}
\value{
A data frame with parameters in columns and simulations in rows. The parameters are returned as factors.
}
\description{
Collect parameters from multiple simulations
}
