% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_sims.R
\name{collect_sims}
\alias{collect_sims}
\title{Collect simulation data}
\usage{
collect_sims(
  root,
  variables,
  by = 1,
  dupl = 1,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  check_extant = TRUE,
  pattern = "sim_",
  level = 0,
  verbose = TRUE,
  pb = TRUE,
  id_column = "sim"
)
}
\arguments{
\item{root}{One of multiple paths to simulation folders or folders into which to recurse to look for simulation folders.}

\item{variables}{Names of the variables to extract}

\item{by, dupl, parnames, combine, as_numeric}{Parameters to be passed to \code{read_data}}

\item{check_extant}{Whether to check for non-extinct and non-crashed simulation folders}

\item{pattern}{Optional pattern to look for if simulation folders are searched by recursion}

\item{level}{Level of recursion. Defaults to 0 for no recursion (hen assumes that \code{root} is a vector of simulation folder paths).}

\item{verbose, pb}{Whether to display messages and progress bars}

\item{id_column}{Optional name of the simulation identifier column}
}
\value{
A data frame
}
\description{
Collect specified variables and parameters from across multiple simulation folders
}
