% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_status.R
\name{collect_status}
\alias{collect_status}
\title{Collect SLURM exit status}
\usage{
collect_status(simulations, pattern = "^sim_", verbose = TRUE, level = 0)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation
folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{level}{Recursion level}
}
\value{
A character vector
}
\description{
Collect SLURM exit status
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

# Should return not found with the example data because these simulations
# were run locally, not on SLURM
collect_status(root, pattern = "example", level = 1)

}

}
