% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Collect data from across simulations}
\usage{
combine_data(roots, type = "data", id_column = "sim", archfiles = NULL, ...)
}
\arguments{
\item{roots}{Paths to simulation folders}

\item{type}{The \code{type} argument of \code{read_this}}

\item{id_column}{Optional name of the simulation identifier column}

\item{archfiles}{Optional vector of architecture file names, each to be passed with their respective simulation path to the \code{read_this} function. Must be the same length as \code{roots} (one architecture for each simulation). The architecture file will be searched for in the simulation folder, so avoid full paths. Only makes sens if \code{type} is "genome" or "network", ignored otherwise.}

\item{...}{Parameters to be passed to \code{read_this}}
}
\value{
A tibble
}
\description{
Collect specified variables and parameters from across multiple simulation
folders
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

# Collect simulation data with some parameters
combine_data(
  root, variables = c("time", "EI"), pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry")
)

combine_data(
  root, variables = "genome_Fst", pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry"), architecture = TRUE, type = "genome"
)

combine_data(
  root, variables = "network_corbreed", pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry"), architecture = TRUE, type = "network"
)

}

}
