% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_genome_heatmap.R
\name{dplot_genome_heatmap}
\alias{dplot_genome_heatmap}
\title{Diagnostic genome heatmap through time}
\usage{
dplot_genome_heatmap(root, y, archfile = "architecture.txt")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to plot.}

\item{archfile}{The architecture file to look for. Defaults to "architecture.txt"}
}
\value{
A ggplot
}
\description{
Quick-plots a given variable across the genome and through time
}
