% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_genome_lines.R
\name{dplot_genome_lines}
\alias{dplot_genome_lines}
\title{Plot locus-specific lines through time}
\usage{
dplot_genome_lines(root, y, span = 0.2)
}
\arguments{
\item{y}{Variable to plot}

\item{span}{Optional span parameter for LOESS smoothing (no smoothing if 0)}

\item{roooooooot}{Path to the simulation}
}
\value{
A ggplot
}
\description{
Plot locus-specific lines through time
}
