% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_genome_ridges.R
\name{dplot_genome_ridges}
\alias{dplot_genome_ridges}
\title{Genome-wide distributions through time}
\usage{
dplot_genome_ridges(root, y, times = NULL)
}
\arguments{
\item{root}{Path to the simulation}

\item{y}{Variable to show}

\item{times}{Optional vector of time points to show}
}
\value{
A ggplot
}
\description{
Genome-wide distributions through time
}
