% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_genome_scan.R
\name{dplot_genome_scan}
\alias{dplot_genome_scan}
\title{Diagnostic genome scan}
\usage{
dplot_genome_scan(root, y, x = "locus", t = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to plot}

\item{x}{Optional x-axis, defaults to locus identity but could also be e.g. "location" for locus position along the genome}

\item{t}{Time point at which to show the scan. If unspecified, last time point.}
}
\value{
A ggplot
}
\description{
Quick-plots a genome scan of a given variable at a given time point
}
