% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_network.R
\name{dplot_network}
\alias{dplot_network}
\title{Plot gene network}
\usage{
dplot_network(root, y, t = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to map as node size aesthetics}

\item{t}{What time point? Defaults to last time point}
}
\value{
A ggplot
}
\description{
Plot gene network
}
