% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_population_density.R
\name{dplot_population_density}
\alias{dplot_population_density}
\title{Diagnose density of a given variable across the population}
\usage{
dplot_population_density(root, y, by = 1, j = 1, t = NULL, fill = "lightgrey")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{by}{How many columns to split the variable into? See \code{?read_data}.}

\item{j}{If the variable is splitted, which column to show?}

\item{t}{Optional time point to filter. Defaults to the last time point}

\item{fill}{Optional fill color}
}
\value{
A ggplot
}
\description{
Diagnose density of a given variable across the population
}
