% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_simulation_line.R
\name{dplot_simulation_line}
\alias{dplot_simulation_line}
\title{Diagnose a simulation variable}
\usage{
dplot_simulation_line(root, y, x = "time", by = 1, j = 1, color = "black")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to track}

\item{x}{Optional variable for the horizontal axis. Defaults to "time".}

\item{j}{If the variable is splitted, which column to show?}

\item{color}{Optional color of the line}

\item{bins}{Number of bins for \code{geom_bin2d}}
}
\value{
A ggplot
}
\description{
Diagnose a simulation variable
}
