% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facettize.R
\name{facettize}
\alias{facettize}
\title{Flexible plot facetting}
\usage{
facettize(
  p,
  facet_rows = NULL,
  facet_cols = NULL,
  facet_wrapped = NULL,
  label_facets = FALSE,
  facet_prefixes = NULL,
  sep = " = "
)
}
\arguments{
\item{facet_rows}{Optional parameter name(s) to facet the plot by rows}

\item{facet_cols}{Optional parameter name(s) to facet the plot by columns}

\item{facet_wrapped}{Whether to automatically fill the facets by row with one-dimensional array of parameter combinations. If TRUE, combines the parameters in facet_rows and facet_cols into a single array. Ignored if none of those is specified.}

\item{label_facets}{Whether to add custom labels to the facets}

\item{facet_prefixes}{Optional prefixes to add to the facet labels of each parameter. Must have one value per facetting parameter. Ignored if label_facets is FALSE. If not specified and label_facets is TRUE, the names of the parameters are used as prefixes.}

\item{sep}{Optional separator to use when adding prefixes. Defaults to " = ".}
}
\value{
A facetted ggplot
}
\description{
Customizable way to split a ggplot into multiple facets depending on facetting variables.
}
