% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dirs.R
\name{fetch_dirs}
\alias{fetch_dirs}
\title{Fetch directories within directories}
\usage{
fetch_dirs(roots, pattern = "^.*$", level = 0)
}
\arguments{
\item{roots}{One or several directories to search in}

\item{pattern}{Optional pattern to look for}

\item{level}{Recursion level}
}
\value{
A vector of directory full names
}
\description{
Fetch directories within directories
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

fetch_dirs(root, pattern = "example", level = 1)

}

}
