% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_completed.R
\name{find_completed}
\alias{find_completed}
\title{Find completed simulations}
\usage{
find_completed(sims)
}
\arguments{
\item{sims}{A vector of simulation folders}
}
\value{
A character vector with the paths to completed simulations
}
\description{
Find completed simulations
}
\details{
Finds completed simulations by checking the SLURM exit status.
Applicable only to simulations run on a server with SLURM.
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

roots <- fetch_dirs(roots, pattern = "example_", level = 1)
find_completed(roots)

}

}
