% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_edge.R
\name{find_edge}
\alias{find_edge}
\title{Find an edge in a gene network}
\usage{
find_edge(i, j, arch)
}
\arguments{
\item{i, j}{The indices of the loci connected by the edge}

\item{arch}{A genetic architecture dataset, on a per-edge basis (see e.g. \code{?read_arch_network})}
}
\value{
A logical vector
}
\description{
Returns a logical vector indicating which is the target edge, from a series of edges in a genetic architecture
}
