% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extinct.R
\name{find_extinct}
\alias{find_extinct}
\title{Find extinct simulations}
\usage{
find_extinct(simulations, pattern = "^sim_", pb = TRUE)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{pb}{Whether to display a progress bar}
}
\description{
Find extinct simulations
}
