% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing.R
\name{find_missing}
\alias{find_missing}
\title{Find missing simulations}
\usage{
find_missing(root, pb = TRUE)
}
\arguments{
\item{root}{Path to the root directory of the simulations to check}

\item{pb}{Whether to display a progress bar}
}
\description{
Find missing simulations
}
