% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_partners.R
\name{find_partners}
\alias{find_partners}
\title{Find the partners of a locus in a network}
\usage{
find_partners(locus, arch)
}
\arguments{
\item{locus}{The focal locus}

\item{arch}{The network-side of the architecture (second list element
returned by \code{read_arch_network}, with \code{as_list} set to TRUE)}
}
\value{
A vector of integers
}
\description{
Find the partners of a locus in a network
}
