% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degrees.R
\name{get_degrees}
\alias{get_degrees}
\title{Get locus degrees}
\usage{
get_degrees(arch)
}
\arguments{
\item{arch}{A genetic architecture}
}
\value{
A vector of degrees across loci
}
\description{
Get locus degrees
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
arch <- read_architecture(root)
get_degrees(arch)

}
