% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_edges_coordinates.R
\name{give_edges_coordinates}
\alias{give_edges_coordinates}
\title{Attach locus coordinates to edges}
\usage{
give_edges_coordinates(data, edges, x, y)
}
\arguments{
\item{data}{A locus-wise data frame. It needs at least a column "locus", and columns named after arguments \code{x} and \code{y}.}

\item{edges}{An edge-wise data frame, with at least a column "from"and a column "to".}

\item{x, y}{Names of the columns of \code{data} that contain locus coordinates.}
}
\value{
An edge-wise data frame
}
\description{
Add x and y corrdinates to each locus of each edge in an edge-wise
dataset.
}
