% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_nedges.R
\name{guess_nedges}
\alias{guess_nedges}
\title{Guess the number of edges}
\usage{
guess_nedges(folder, variable = "network_corfreq")
}
\arguments{
\item{folder}{Path to the folder}

\item{variable}{The edge-specific variable to use to guess the number of
edges}
}
\value{
An integer
}
\description{
Function to find out the number of edges in a given simulation
}
\details{
The file \code{time.dat} must be present
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

guess_nedges(root)

}

}
