% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_nrows.R
\name{guess_nrows}
\alias{guess_nrows}
\title{Guess the number of rows per time point}
\usage{
guess_nrows(folder, variable)
}
\arguments{
\item{folder}{Path to the folder}

\item{variable}{The variable to use to guess the number of rows}
}
\value{
An integer
}
\description{
Function to find out the number of rows per time point in a given simulation
}
\details{
The file \code{time.dat} must be present
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

guess_nrows(root, "EI")

}

}
