% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete.R
\name{is_complete}
\alias{is_complete}
\title{Is a simulation completed?}
\usage{
is_complete(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A logical
}
\description{
Checks for the word "COMPLETED" in the SLURM output
}
\examples{

\dontrun{

# Should return an error because examples were not run using SLURM
is_complete("data/example_1")

}

}
