% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_extant.R
\name{is_extant}
\alias{is_extant}
\title{Is a simulation extant?}
\usage{
is_extant(root, logfile = "^log.txt$", slurm = FALSE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{logfile}{Optional log file name}

\item{slurm}{Whether to check SLURM output files}
}
\value{
A logical
}
\description{
Checks that the simulation is non-missing, succesfully completed and did
not go extinct.
}
\examples{

\dontrun{

is_extant("data/example_1", slurm = FALSE)

}

}
