% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_extinct.R
\name{is_extinct}
\alias{is_extinct}
\title{Did a simulation go extinct?}
\usage{
is_extinct(folder, logfile = "log.txt", slurm = FALSE)
}
\arguments{
\item{folder}{Path to the folder}

\item{logfile}{Name of the log file to read}

\item{slurm}{Whether to look up in the SLURM output file (instead of
\code{logfile})}
}
\value{
A logical
}
\description{
Did a simulation go extinct?
}
\details{
This function looks for the word "extinct" in the log file of the
simulation
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

is_extinct(root)

}

}
