% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_missing.R
\name{is_missing}
\alias{is_missing}
\title{Is a simulation missing?}
\usage{
is_missing(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A logical
}
\description{
Checks for the presence of \code{.dat} files
}
\examples{

\dontrun{

is_missing("data/example_1")

}

}
