% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_facet_labels.R
\name{make_facet_labels}
\alias{make_facet_labels}
\title{Make facet labels}
\usage{
make_facet_labels(data, varname)
}
\arguments{
\item{data}{The dataset}

\item{varname}{The variable used for facetting}
}
\description{
Make facet labels of the form "variable = value".
}
