% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrep.R
\name{mrep}
\alias{mrep}
\title{Repeat many elements many times}
\usage{
mrep(x, n)
}
\arguments{
\item{x}{A vector of things to repeat}

\item{n}{A vector of how many times to repeat each thing}
}
\value{
A vector of the same type as \code{x}
}
\description{
Repeat many elements many times
}
