% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_data.R
\name{pivot_data}
\alias{pivot_data}
\title{Pivot data to long format}
\usage{
pivot_data(data, variables, newnames = NULL)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The columns to pass as \code{cols} in \code{tidyr::pivot_longer}}

\item{newnames}{Optional character vector of new names for \code{variables}}
}
\value{
A data frame
}
\description{
Can be useful when plotting with facets.
}
\examples{

\dontrun{

# Location of the simulation data
root <- "data/example_1"

# Read speciation metrics in the long format
data <- read_sim(root, c("EI", "RI", "SI"))
pivot_data(data, c("EI", "RI", "SI"))

# Read individual trait values in the long format
data <- read_pop(root, "individual_trait", by = 3)
cols <- paste0("individual_trait", 1:3)
pivot_data(data, cols, newnames = 0:2)

}

}
