% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_bin2d.R
\name{plot_genome_bin2d}
\alias{plot_genome_bin2d}
\title{Plot genome distribution through time}
\usage{
plot_genome_bin2d(
  root,
  y,
  x = "time",
  t = NULL,
  facet_by = NULL,
  archfile = "architecture.txt",
  ...
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{x}{Optional variable for the horizontal axis. Defaults to "time".}

\item{t}{What time point? Defaults to last time point, only if \code{x} is not "time"}

\item{facet_by}{Optional variable to facet by}

\item{archfile}{Optional architecture file name}

\item{...}{Parameters to be passed to \code{geom_path}}
}
\value{
A ggplot
}
\description{
Plot genome distribution through time
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_genome_bin2d(root, "genome_Fst")

# Plot Fst versus Cst
plot_genome_bin2d(root, "genome_Fst", x = "genome_Cst")

# Can facet by some variable
plot_genome_bin2d(root, "genome_Fst", facet_by = "trait")

}
