% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_density.R
\name{plot_genome_density}
\alias{plot_genome_density}
\title{Plot genomic distribution}
\usage{
plot_genome_density(
  root,
  y,
  plot_type = "violin",
  t = NULL,
  archfile = "architecture.txt",
  ...
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to plot}

\item{plot_type}{See \code{?ggsim::ggdensityplot}}

\item{t}{Optional time point. Last time point if unspecified.}

\item{archfile}{Optional architecture file name}

\item{...}{Parameters to be passed to \code{ggsim::ggdensityplot}}
}
\value{
A ggplot
}
\description{
Plot genomic distribution
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_genome_density(root, "genome_Fst", x = "trait")

}
