% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_distributions.R
\name{plot_genome_distributions}
\alias{plot_genome_distributions}
\title{Plot distributions across loci per trait}
\usage{
plot_genome_distributions(
  data,
  y,
  t = NULL,
  timed = TRUE,
  jitter = TRUE,
  width = 0.2
)
}
\arguments{
\item{data}{A locus-wise dataset.}

\item{y}{Name of the variable to plot.}

\item{t}{Optional time point to filter. Defaults to the last one if \code{timed} is TRUE.}

\item{timed}{Whether there are multiple time points in the dataset (the column must be named "time").}

\item{jitter}{Logical. Whether to show jittered points.}

\item{width}{Width of the jitter.}
}
\value{
A ggplot
}
\description{
Plot distributions across loci per trait
}
