% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_heatmap.R
\name{plot_genome_heatmap}
\alias{plot_genome_heatmap}
\title{Plot genomic heatmap through time}
\usage{
plot_genome_heatmap(data, z)
}
\arguments{
\item{data}{A locus-wise data frame with columns "time", "locus" and \code{z} at least}

\item{z}{Name of the variable to plot}
}
\value{
A ggplot
}
\description{
Heatmap of a variable through time across the genome.
}
