% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_ridges.R
\name{plot_genome_ridges}
\alias{plot_genome_ridges}
\title{Plot genome distribution through time}
\usage{
plot_genome_ridges(
  root,
  y,
  times = NULL,
  facet_by = NULL,
  archfile = "architecture.txt"
)
}
\arguments{
\item{root}{Path to the simulation}

\item{y}{Variable to show}

\item{times}{Optional vector of time points to show}

\item{facet_by}{Variable to facet by}

\item{archfile}{Optional architecture file name}
}
\value{
A ggplot
}
\description{
Plot genome distribution through time
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_genome_ridges(root, "genome_Fst", times = c(100, 200, 300))
plot_genome_ridges(
 root, "genome_Fst", times = c(100, 200, 300), facet_by = "trait"
)

}
