% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot gene network}
\usage{
plot_network(root, y, t = NULL, archfile = "architecture.txt")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to map as node size aesthetics}

\item{t}{What time point? Defaults to last time point}

\item{archfile}{Optional architecture file name}
}
\value{
A ggplot
}
\description{
Plot gene network
}
