% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_blank.R
\name{plot_network_blank}
\alias{plot_network_blank}
\title{Plot gene network without evolving variable}
\usage{
plot_network_blank(root, y = NULL, archfile = "architecture.txt")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Optional name of the variable to map as node size aesthetics.
Should be in the fields returned by \code{read_genome_architecture}.}

\item{archfile}{Optional architecture file name}
}
\value{
A ggplot
}
\description{
Plot gene network without evolving variable
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_network_blank(root, y = "degree")

}
