% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_bin2d.R
\name{plot_pop_bin2d}
\alias{plot_pop_bin2d}
\title{Plot population distribution through time}
\usage{
plot_pop_bin2d(root, y, x = "time", by = 1, j = 1, k = 1, t = NULL, ...)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{x}{Optional variable for the horizontal axis. Defaults to "time".}

\item{by}{Numbers of columns in which to split variable \code{x} (only if \code{x} is
not "time" and not \code{y}) and \code{y}. See \code{?read_data}.}

\item{j}{If variable \code{y} is splitted, which column to show?}

\item{k}{If variable \code{x} is splitted, which column to show?}

\item{t}{What time point? Defaults to last time point, only if \code{x} is not
"time"}

\item{...}{Parameters to be passed to \code{geom_path}}
}
\value{
A ggplot
}
\description{
Plot population distribution through time
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_pop_bin2d(root, "individual_trait", by = 3, j = 1)

# Plot the mating versus ecological trait
plot_pop_bin2d(
  root, "individual_trait", x = "individual_trait", by = 3,
  j = 2, k = 1
)

}
