% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_density.R
\name{plot_pop_density}
\alias{plot_pop_density}
\title{Plot population distribution}
\usage{
plot_pop_density(root, y, by = 1, j = 1, plot_type = "density", t = NULL, ...)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{by}{How many columns to split the variable into? See \code{?read_data}.}

\item{j}{If the variable is splitted, which column to show?}

\item{plot_type}{See \code{?ggsim::ggdensityplot}}

\item{t}{Optional time point to filter. Defaults to the last time point}

\item{...}{Parameters to be passed to \code{ggsim::ggdensityplot}}

\item{fill}{Optional fill color}
}
\value{
A ggplot
}
\description{
Plot population distribution
}
\examples{

# Plot the distribution of the ecological trait
root <- system.file("extdata", "example_1", package = "egssimtools")
plot_pop_density(root, "individual_trait", by = 3, j = 1)

}
