% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_ridges.R
\name{plot_pop_ridges}
\alias{plot_pop_ridges}
\title{Plot population distribution through time}
\usage{
plot_pop_ridges(
  root,
  y,
  by = 1,
  j = 1,
  times = NULL,
  facet_by = NULL,
  cpp_numbering = TRUE,
  ...
)
}
\arguments{
\item{root}{Path to the simulation}

\item{y}{Variable to show}

\item{by}{Numbers of columns in which to split variable \code{y}. See \code{?read_data}.}

\item{j}{If variable \code{y} is splitted, which column to show?}

\item{times}{Optional vector of time points to show}

\item{facet_by}{Variable to facet by}

\item{cpp_numbering}{Parameter for \code{read_indiv_long}. Relevant only if
\code{facet_by} is specified.}

\item{...}{Parameters to be passed to \code{ggridges::geom_density_ridges}}
}
\value{
A ggplot
}
\description{
Plot population distribution through time
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")

# Plot the distribution of the ecological trait at three time points
plot_pop_ridges(
  root, "individual_trait", by = 3, j = 1, times = c(100, 200, 300)
)

# Plot the distribution for the three traits
plot_pop_ridges(
  root, "individual_trait", by = 3, j = 1, times = c(100, 200, 300),
  facet_by = "trait"
)

}
