% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population_bin2d.R
\name{plot_population_bin2d}
\alias{plot_population_bin2d}
\title{Diagnose density of a given variable across the population through time}
\usage{
plot_population_bin2d(root, y, by = 1, j = 1, bins = 50)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{by}{How many columns to split the variable into? See \code{?read_data}.}

\item{j}{If the variable is splitted, which column to show?}

\item{bins}{Number of bins for \code{geom_bin2d}}
}
\value{
A ggplot
}
\description{
Diagnose density of a given variable across the population through time
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_population_bin2d(root, "individual_trait", by = 3, j = 1)

}
