% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population_density.R
\name{plot_population_density}
\alias{plot_population_density}
\title{Diagnose density of a given variable across the population}
\usage{
plot_population_density(root, y, by = 1, j = 1, t = NULL, fill = "lightgrey")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot the density for}

\item{by}{How many columns to split the variable into? See \code{?read_data}.}

\item{j}{If the variable is splitted, which column to show?}

\item{t}{Optional time point to filter. Defaults to the last time point}

\item{fill}{Optional fill color}
}
\value{
A ggplot
}
\description{
Diagnose density of a given variable across the population
}
\examples{

# Plot the distribution of the ecological trait
root <- system.file("extdata", "example_1", package = "egssimtools")
plot_population_density(root, "individual_trait", by = 3, j = 1)

}
