% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_line.R
\name{plot_sim_line}
\alias{plot_sim_line}
\title{Plot a simulation-wise variable}
\usage{
plot_sim_line(root, y, x = "time", by = NULL, j = 1, k = 1, ...)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to plot}

\item{x}{Optional variable for the horizontal axis. Defaults to "time".}

\item{by}{Numbers of columns in which to split variable \code{x} (only if \code{x} is not "time") and \code{y}. See \code{?read_data}.}

\item{j}{If variable \code{y} is splitted, which column to show?}

\item{k}{If variable \code{x} is splitted, which column to show?}

\item{...}{Parameters to be passed to \code{geom_path}}
}
\value{
A ggplot
}
\description{
Plot a simulation-wise variable
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_sim_line(root, "EI")
plot_sim_line(root, "RI", x = "EI")

}
