% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulations.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Plot multiple simulations through time}
\usage{
plot_simulations(
  root,
  variable,
  pattern = "^sim_",
  color_by = NULL,
  color_by_numeric = TRUE,
  colors = NULL,
  pb = TRUE,
  verbose = TRUE,
  facet_rows = NULL,
  facet_cols = NULL,
  facet_wrapped = FALSE,
  reverse_order = NULL,
  label_facets = FALSE,
  facet_prefixes = NULL,
  sep = " = "
)
}
\arguments{
\item{root}{Where the simulation folders are}

\item{variable}{What variable to plot}

\item{pattern}{Optional pattern characteristic of simulation folders. Defaults to starting with "sim_".}

\item{color_by}{Optional parameter to use to color the lines}

\item{color_by_numeric}{Whether the color parameter must be treated as numeric and be mapped to a color gradient. Defaults to TRUE. If FALSE, it is treated as a factor.}

\item{colors}{Optional colors of the lines. If color_by is NULL, the color that all the lines must have. If color_by is defined and color_by_numeric is TRUE, the lower and the upper end of the color gradient. If color_by_numeric is FALSE, a set of colors for each level of the color factor. If not defined, default ggplot color(s) will be used.}

\item{pb}{Whether to display progress bars}

\item{verbose}{Whether to display messages}

\item{facet_rows}{Optional parameter name(s) to facet the plot by rows}

\item{facet_cols}{Optional parameter name(s) to facet the plot by columns}

\item{facet_wrapped}{Whether to automatically fill the facets by row with one-dimensional array of parameter combinations. If TRUE, combines the parameters in facet_rows and facet_cols into a single array. Ignored if none of those is specified.}

\item{reverse_order}{Optional name of a parameter for which to reverse the order of the facets (e.g. to put highest values on the top rows)}

\item{label_facets}{Whether to add custom labels to the facets}

\item{facet_prefixes}{Optional prefixes to add to the facet labels of each parameter. Must have one value per facetting parameter. Ignored if label_facets is FALSE. If not specified and label_facets is TRUE, the names of the parameters are used as prefixes.}

\item{sep}{Optional separator to use when adding prefixes. Defaults to " = ".}
}
\value{
A plot showing a variable through time for multiple simulations.
}
\description{
This function allows to easily loop through many simulation folders, extract data from them and plot them synthetically. Some population-wide variables are plotted through time for each simulation.
}
\note{
The lines have different transparency levels, all very close to 0.5. This is a hack to make sure that all simulations can efficiently be plotted together on the same plot.
}
