% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_density.R
\name{plot_trait_density}
\alias{plot_trait_density}
\title{Heatmap of trait values through time}
\usage{
plot_trait_density(folder, trait, bins = 15)
}
\arguments{
\item{trait}{What trait to plot? (1) Ecological trait, (2) mating preference and (3) neutral trait.}

\item{bins}{Number of bins in the heatmap}

\item{filename}{Path to the folder}
}
\value{
A ggplot with the basic needed components. You can add extra layers to it (e.g. change the axis labels) outside the function.
}
\description{
Shows the densities of trait values through time.
}
