% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arch_file.R
\name{read_arch_file}
\alias{read_arch_file}
\title{Read an architecture file}
\usage{
read_arch_file(filename)
}
\arguments{
\item{filename}{Path to the architecture file}
}
\value{
A list with all the fields of a genetic architecture
}
\description{
Read an architecture file
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

f <- file.path(root, "architecture.txt")
read_arch_file(f)

}

}
