% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arch_genome.R
\name{read_arch_genome}
\alias{read_arch_genome}
\title{Read locus-specific genome architecture}
\usage{
read_arch_genome(folder, filename = "architecture.txt")
}
\arguments{
\item{folder}{Path to the simulation}

\item{filename}{Name of the architecture file}
}
\value{
A tibble with, for each locus, its location, trait, effect,
dominance, chromosome and degree
}
\description{
Make a table from locus-specific genetic architecture details
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

read_arch_genome(root)

}

}
