% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arch_network.R
\name{read_arch_network}
\alias{read_arch_network}
\title{Read gene network architecture}
\usage{
read_arch_network(folder, filename = "architecture.txt", as_list = FALSE)
}
\arguments{
\item{folder}{Path to the simulation}

\item{filename}{Optional architecture file name}

\item{as_list}{Whether to return the output at a list of two data frames
instead of a \code{tbl_graph}.}
}
\value{
A \code{tbl_graph} object (useful for plotting using the \code{ggraph}
package), or a list of data frames if applicable
}
\description{
Reads a genetic architecture and stores it into a \code{tbl_graph} object
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

read_arch_network(root)

}

}
