% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_architecture.R
\name{read_architecture}
\alias{read_architecture}
\title{Read architecture}
\usage{
read_architecture(folder, filename = "architecture.txt")
}
\arguments{
\item{folder}{Path to the folder}

\item{filename}{Optional file name}
}
\description{
Read architecture
}
