% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bitset.R
\name{read_bitset}
\alias{read_bitset}
\title{Read bitset}
\usage{
read_bitset(filename)
}
\arguments{
\item{filename}{Path to the file to read}
}
\value{
A vector if integers
}
\description{
Read a binary file bit-wise and return a vector of integers (0/1)
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

f <- file.path(root, "time.dat")
read_bitset(f)

}

}
