% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read simulation data}
\usage{
read_data(
  folder,
  variables,
  by = 1,
  dupl = 1,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  architecture = FALSE,
  archfile = "architecture.txt",
  parfile = "paramlog.txt"
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read}

\item{by}{A list. For each variable, length of the chunks to split it by. Recycled if its length is one.}

\item{dupl}{A list. For each variable, how many times to duplicate each row. Provide an integer to duplicate every row the same number of times, a vector of integers to duplicate each row a specific number of times, or a character string to read a vector of number of times from a .dat file. This argument is recycled if its length is one.}

\item{parnames}{Character vector of parameter names to append to the simulation data}

\item{combine, as_numeric}{Parameters for \code{read_parameters}}

\item{architecture}{Whether to read and append locus-wise genetic architecture. Make sure then that you are reading \code{variables} that are locus-specific variables in long-format (one column per variable).}

\item{archfile}{Name of the architecture file, if needed}

\item{parfile}{Name of the parameter file, if needed}
}
\value{
A data frame
}
\description{
Function to read simulation data from binary files into a data frame. Can read multiple data files and bind them together.
}
\note{
Do not provide the extension of the files. It is assumed to be \code{.dat}.
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
read_data(root, c("time", "genome_Fst"), by = c(1, 300))

}
