% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indiv.R
\name{read_indiv}
\alias{read_indiv}
\title{Read individual data through time}
\usage{
read_indiv(
  folder,
  variables,
  by = 3,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  parfile = "paramlog.txt"
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{by}{Same as the \code{by} argument of \code{read_data} (do not count \code{time} in)}

\item{parnames, combine, as_numeric, parfile}{Parameters for \code{read_data}}
}
\value{
A data frame
}
\description{
Wrapper around \code{read_data} to read individual-wise data
}
\details{
The files \code{time.dat} and \code{population_size.dat} must be present
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
read_indiv(root, "individual_trait", by = 3)

}
