% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indiv_loci.R
\name{read_indiv_loci}
\alias{read_indiv_loci}
\title{Read individual loci}
\usage{
read_indiv_loci(
  root,
  nloci = NULL,
  t = NULL,
  extra = NULL,
  freezerfile = "freezer.dat",
  locifile = "locivalues.dat"
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{nloci}{Number of loci. Guessed using \code{guess_nloci} if NULL.}

\item{t}{Optional time points to read. All if NULL.}

\item{extra}{Optional names of extra individual-wise variables to attach to
the output data frame (read using \code{read_pop})}

\item{freezerfile}{File name of the freezer file}

\item{locifile}{File name of the loci values file}
}
\value{
A data frame
}
\description{
Makes a data frame with alleles at each locus for each individual.
Can be very large
}
\examples{

\dontrun{

root <- "data/example_1"

read_indiv_loci(root)

}

}
