% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indiv_long.R
\name{read_indiv_long}
\alias{read_indiv_long}
\title{Read individual data through time (long format)}
\usage{
read_indiv_long(
  folder,
  variables,
  by = 3,
  to_long = variables[1],
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  parfile = "paramlog.txt",
  names_to = "variable",
  replace_with = "",
  cpp_numbering = TRUE
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{by}{See \code{?read_indiv}.}

\item{to_long}{What variable (within \code{variables}) to use to pivot the data
in the long format? Defaults to the first element in \code{variables}.
See \code{tidyr::pivot_longer} for details.}

\item{parnames, combine, as_numeric, parfile}{Parameters for \code{read_data}}

\item{names_to}{Optional name for the new column containing variable names.
See \code{?tidyr::pivot_longer} for details.}

\item{replace_with}{Optional string to replace \code{to_long} in the column that
will contain the variable names (e.g. change from "individual_trait1" to
"trait 1"). No replacement if NULL.}

\item{cpp_numbering}{Whether to keep the C++ numbering style in variable
names (starting from 0)}
}
\value{
A data frame
}
\description{
Wrapper around \code{read_data} to read individual-wise data in the long format
(useful for facetting plots).
}
\details{
The files \code{time.dat} and \code{population_size.dat} must be present.
Multiple variables can be read, but one has to be chosen for pivoting to the
long format.
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
read_indiv_long(root, "individual_trait", by = 3, replace_with = "trait ")

}
