% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_individuals.R
\name{read_individuals}
\alias{read_individuals}
\title{Read individual data}
\usage{
read_individuals(folder, variable)
}
\arguments{
\item{variable}{What to read}

\item{filename}{Path to the folder}
}
\description{
Read individual data
}
