% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loci.R
\name{read_loci}
\alias{read_loci}
\title{Read locus-specific data through time}
\usage{
read_loci(
  folder,
  variables,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  architecture = FALSE,
  archfile = "architecture.txt",
  parfile = "paramlog.txt"
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{parnames, combine, as_numeric, architecture, archfile, parfile}{Parameters for \code{read_data}}
}
\value{
A data frame
}
\description{
Wrapper around \code{read_data} to read locus-wise data
}
\details{
The file \code{time.dat} must be present
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
read_loci(root, "genome_Fst")

}
