% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_network.R
\name{read_network}
\alias{read_network}
\title{Read edge-specific data through time}
\usage{
read_network(
  folder,
  variables,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  architecture = FALSE,
  archfile = "architecture.txt",
  parfile = "paramlog.txt",
  nedges = NULL
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{parnames, combine, as_numeric, parfile}{Parameters
for \code{read_data}}

\item{architecture}{Whether to attach genetic architecture data}

\item{archfile}{Optional name of the genetic architecture file}

\item{nedges}{Number of edges (automatically guessed if unspecified)}
}
\value{
A tibble
}
\description{
Wrapper around \code{read_data} to read edge-wise data
}
\details{
The file \code{time.dat} must be present
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

# Read Fst throughout the genome
read_edges(root, "network_corbreed")

# Read multiple metrics and attach architecture
variables <- paste0("network_", c("corbreed", "corfreq"))
read_network(root, variables, architecture = TRUE)

}

}
