% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_network_architecture.R
\name{read_network_architecture}
\alias{read_network_architecture}
\title{Read gene network architecture}
\usage{
read_network_architecture(folder, filename = "architecture.txt")
}
\arguments{
\item{folder}{Path to the simulation}

\item{filename}{Optional architecture file name}
}
\value{
A \code{tbl_graph} object
}
\description{
Reads a genetic architecture and stores it into a \code{tbl_graph} object
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
read_network_architecture(root)

}
