% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param.R
\name{read_param}
\alias{read_param}
\title{Read parameters}
\usage{
read_param(
  folder,
  parnames = NULL,
  filename = "paramlog.txt",
  combine = FALSE,
  flatten = FALSE,
  as_numeric = NULL
)
}
\arguments{
\item{folder}{Path to the folder}

\item{parnames}{Optional vector of parameter names}

\item{filename}{Optional file name}

\item{combine}{Whether to paste together compound parameters}

\item{flatten}{Whether to return a list of single values. If FALSE, some of
the elements may be vectors of multiple values (compound parameters).}

\item{as_numeric}{Which parameters to convert to numeric?}
}
\value{
A list of parameter values
}
\description{
Read parameters
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

read_param(root)

}

}
