% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param_file.R
\name{read_param_file}
\alias{read_param_file}
\title{Read a parameter file}
\usage{
read_param_file(filename, parnames = NULL)
}
\arguments{
\item{filename}{Path to the parameter file}

\item{parnames}{Optional vector of parameter names}
}
\value{
A list of parameter values
}
\description{
Read a parameter file
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

f <- file.path(root, "parameters.txt")
read_param_file(f)

}

}
