% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read parameters}
\usage{
read_parameters(folder, parnames = NULL, filename = "paramlog.txt")
}
\arguments{
\item{folder}{Path to the folder}

\item{parnames}{Optional vector of parameter names}

\item{filename}{Optional file name}
}
\description{
Read parameters
}
