% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paramfile.R
\name{read_paramfile}
\alias{read_paramfile}
\title{Read a parameter file}
\usage{
read_paramfile(filename, parnames = NULL)
}
\arguments{
\item{filename}{Path to the parameter file}

\item{parnames}{Optional vector of parameter names}
}
\value{
A list of parameter values
}
\description{
Read a parameter file
}
