% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pedigree.R
\name{read_pedigree}
\alias{read_pedigree}
\title{Read the results of a pedigree experiment}
\usage{
read_pedigree(root, k, tped, tend, filename = "pedigree.dat")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{k}{Number of offspring per cross}

\item{tped}{Frequency of the pedigree experiment, in generations}

\item{tend}{Number of generations in the simulation}

\item{filename}{Pedigree file name}
}
\value{
A data frame with information, for each cross, about the male and the female, their ecotypes, their trait values and the trait values of their k offspring. There is an extra column for time point in case the experiment was conducted at multiple time points.
}
\description{
The pedigree experiment saves distributions of offspring of random mating
pairs
}
