% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population.R
\name{read_population}
\alias{read_population}
\title{Read population-wide data}
\usage{
read_population(folder, variable)
}
\arguments{
\item{variable}{What to read}

\item{filename}{Path to the folder}
}
\description{
Read population-wide data
}
