% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population_size.R
\name{read_population_size}
\alias{read_population_size}
\title{Read total population size}
\usage{
read_population_size(folder)
}
\arguments{
\item{filename}{Path to the folder}
}
\description{
Read total population size
}
