% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_resources.R
\name{read_resources}
\alias{read_resources}
\title{Read resource concentrations}
\usage{
read_resources(folder)
}
\arguments{
\item{filename}{Path to the folder}
}
\description{
Equilibrium resource concentrations by resource by habitat by time point.
}
