% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sim.R
\name{read_sim}
\alias{read_sim}
\title{Read simulation summary through time}
\usage{
read_sim(
  folder,
  variables,
  by = 1,
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  parfile = "paramlog.txt"
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{by}{Same as the \code{by} argument of \code{read_data} (do not count \code{time} in)}

\item{parnames, combine, as_numeric, parfile}{Parameters for \code{read_data}}
}
\value{
A data frame
}
\description{
Wrapper around \code{read_data} to read timepoint-wise summary data
}
\details{
The file \code{time.dat} must be present
}
