% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_this.R
\name{read_this}
\alias{read_this}
\title{Wrapper around data readers}
\usage{
read_this(root, type = "data", ...)
}
\arguments{
\item{root}{Path to a simulation folder}

\item{type}{Which \verb{read_*} function to call, either of \code{data}, \code{sim}, \code{pop},
\code{genome} or \code{network}}

\item{...}{Parameters to be passed to the reading function}
}
\value{
A tibble
}
\description{
This function works as an alias for the \verb{read_*} family of functions
}
\examples{

\dontrun{

root <- "data/example_1"

read_this(root, "sim", variables = "EI")
read_this(root, "data", variables = c("time", "EI"))

}

}
\seealso{
\code{read_data}, \code{read_sim}, \code{read_pop}, \code{read_genome}, \code{read_network}
}
