% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_time_if.R
\name{read_time_if}
\alias{read_time_if}
\title{Read time only if available}
\usage{
read_time_if(folder, ntimes = NULL)
}
\arguments{
\item{ntimes}{Number of time points if the time file is not present}

\item{filename}{Path to the folder}
}
\description{
Read the time.dat file only if it is present. If not, return a sequence of time point indices.
}
