% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_str.R
\name{rename_str}
\alias{rename_str}
\title{Function to rename the columns of a data frame given a character vector}
\usage{
rename_str(df, newcols)
}
\arguments{
\item{df}{A data frame}

\item{newcols}{A character string with the new names of the columns}
}
\value{
A renamed data frame
}
\description{
Function to rename the columns of a data frame given a character vector
}
\examples{

df <- data.frame(a = 1, b = 1)
rename_str(df, c("aa", "bb"))

}
