% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_file.R
\name{set_param_file}
\alias{set_param_file}
\title{Set up parameter file}
\usage{
set_param_file(pars, template = NULL, saveto = NULL)
}
\arguments{
\item{pars}{A named list or vector of parameter values. Numeric values will be coerced to character. Please provide compound parameters as strings with space-separated values. Names should be the names of the parameters.}

\item{template}{Optional file to add or change parameters into (the file is not overwritten).}

\item{saveto}{Optional file where to save the output.}
}
\value{
A vector of parameter entries. Not very useful if \code{saveto} is NULL.
}
\description{
Function to create a parameter file with the requested parameter values
}
