% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothen_data.R
\name{smoothen_data}
\alias{smoothen_data}
\title{Smoothen lines using LOESS}
\usage{
smoothen_data(data, x, y, span = 0.2, line = NULL)
}
\arguments{
\item{data}{A data frame}

\item{x}{Name of the horizontal axis variable}

\item{y}{Name of the vertical axis variable}

\item{span}{Span parameter for the smoothing}

\item{line}{Name of the variable identifying the different lines. Leave
unspecified if there is only one line to smoothen.}
}
\value{
A data frame
}
\description{
Smoothen lines using LOESS
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

data <- read_genome(root, "genome_Fst", architecture = TRUE)
smoothen_data(data, "time", "genome_Fst", line = "locus")

}

}
