% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2vec.R
\name{str2vec}
\alias{str2vec}
\title{String to numeric vector}
\usage{
str2vec(string, sep = " ")
}
\arguments{
\item{string}{A string containing numbers}

\item{sep}{The separator between numbers}
}
\value{
A numeric vector
}
\description{
String to numeric vector
}
