% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_parameters.R
\name{collect_parameters}
\alias{collect_parameters}
\title{Collect parameters from multiple simulations}
\usage{
collect_parameters(
  simulations,
  parnames = NULL,
  pattern = "^sim_",
  verbose = TRUE,
  pb = TRUE,
  filename = "paramlog.txt",
  check_extant = TRUE,
  to_numeric = NULL
)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{parnames}{Optional vector of parameter names (collects all parameters if not specified)}

\item{pattern}{Optional pattern characteristic of simulation folders. Defaults to starting with "sim_".}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display a progress bar}

\item{filename}{Optional file name}

\item{check_extant}{Whether to filter extant simulations. Defaults to TRUE. Set to FALSE if e.g. you supplied a vector of simulations you know did not go missing or extinct.}

\item{to_numeric}{Which parameters to convert into numeric. Because all parameters may not be numbers, they are read as factor by default.}
}
\value{
A data frame with parameters in columns and simulations in rows. The parameters are returned as factors.
}
\description{
Collect parameters from multiple simulations
}
