% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_status.R
\name{collect_status}
\alias{collect_status}
\title{Collect SLURM exit status}
\usage{
collect_status(simulations, pattern = "^sim_", verbose = TRUE, pb = TRUE)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display a progress bar}
}
\description{
Collect SLURM exit status
}
