% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_status.R
\name{collect_status}
\alias{collect_status}
\title{Collect SLURM exit status}
\usage{
collect_status(sims)
}
\arguments{
\item{sims}{A vector of simulation folders}
}
\value{
A character vector
}
\description{
Collect SLURM exit status
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

roots <- fetch_dirs(root, pattern = "example", level = 1)

# Should return not found with the example data because these simulations
# were run locally, not on SLURM
collect_status(roots)

}

}
