% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Collect data from across simulations}
\usage{
combine_data(
  root,
  check_extant = FALSE,
  pattern = "sim_",
  level = 0,
  type = "data",
  id_column = "sim",
  ...
)
}
\arguments{
\item{root}{One of multiple paths to simulation folders or folders into
which to recurse to look for simulation folders.}

\item{check_extant}{Whether to check for non-extinct and non-crashed
simulation folders}

\item{pattern}{Optional pattern to look for if simulation folders are
searched by recursion}

\item{level}{Level of recursion. Defaults to 0 for no recursion (then assumes
that \code{root} is a vector of simulation folder paths).}

\item{type}{The \code{type} argument of \code{read_this}}

\item{id_column}{Optional name of the simulation identifier column}

\item{...}{Parameters to be passed to \code{read_this}}
}
\value{
A tibble
}
\description{
Collect specified variables and parameters from across multiple simulation
folders
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

# Collect simulation data with some parameters
combine_data(
  root, variables = c("time", "EI"), pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry")
)

combine_data(
  root, variables = "genome_Fst", pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry"), architecture = TRUE, type = "genome"
)

combine_data(
  root, variables = "network_corbreed", pattern = "example", level = 1,
  parnames = c("ecosel", "hsymmetry"), architecture = TRUE, type = "network"
)

}

}
