% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_completed.R
\name{find_completed}
\alias{find_completed}
\title{Find completed simulations}
\usage{
find_completed(
  simulations,
  pattern = "^sim_",
  verbose = TRUE,
  pb = TRUE,
  level = 0
)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display a progress bar}

\item{level}{Recursion level}
}
\description{
Find completed simulations
}
