% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_completed.R
\name{find_completed}
\alias{find_completed}
\title{Find completed simulations}
\usage{
find_completed(sims, pattern = "^sim_", verbose = TRUE, level = 0)
}
\arguments{
\item{sims}{Either path to a root directory containing simulation folders,
or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{level}{Recursion level}
}
\value{
A character vector with the paths to completed simulations
}
\description{
Find completed simulations
}
\details{
Finds completed simulations by checking the SLURM exit status.
Applicable only to simulations run on a server with SLURM.
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

find_completed(root, pattern = "example_", level = 1)

}

}
