% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extant.R
\name{find_extant}
\alias{find_extant}
\title{Find extant simulations}
\usage{
find_extant(simulations, pattern = "^sim_", verbose = TRUE, pb = TRUE)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display a progress bar}
}
\value{
A vector of names of the simulation folders that completed
}
\description{
Find extant simulations
}
