% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extant.R
\name{find_extant}
\alias{find_extant}
\title{Find extant simulations}
\usage{
find_extant(roots)
}
\arguments{
\item{roots}{Vector of simulation folders}
}
\value{
A character vector
}
\description{
Find simulations that did not crash, that successfully completed and that did
not go extinct
}
\examples{

\dontrun{

roots <- fetch_dirs("data", "example", 1)
find_extant(roots)

}

}
