% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extant.R
\name{find_extant}
\alias{find_extant}
\title{Find extant simulations}
\usage{
find_extant(sims, ...)
}
\arguments{
\item{sims}{A vector of simulation folders}

\item{...}{Parameters for \code{find_extinct}}
}
\value{
A vector of names of the simulation that did not go extinct
}
\description{
Find extant simulations
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

roots <- fetch_dirs(root, pattern = "example_", level = 1)
find_extant(roots)

}

}
