% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extinct.R
\name{find_extinct}
\alias{find_extinct}
\title{Find extinct simulations}
\usage{
find_extinct(root, pb = TRUE)
}
\arguments{
\item{root}{Path to the root directory of the simulations to check}

\item{pb}{Whether to display a progress bar}
}
\description{
Find extinct simulations
}
