% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extinct.R
\name{find_extinct}
\alias{find_extinct}
\title{Find extinct simulations}
\usage{
find_extinct(sims, ...)
}
\arguments{
\item{sims}{A vector of simulation folders}

\item{...}{Parameters for \code{is_extinct}}
}
\value{
A character vector of extinct simulations, or NULL if no simulation
is extinct
}
\description{
Find extinct simulations
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

roots <- fetch_dirs(root, "example", level = 1)
find_extinct(roots)

}

}
