% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extinct.R
\name{find_extinct}
\alias{find_extinct}
\title{Find extinct simulations}
\usage{
find_extinct(sims, pattern = "^sim_", verbose = TRUE, level = 0)
}
\arguments{
\item{sims}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{level}{Recursion level}
}
\description{
Find extinct simulations
}
