% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing.R
\name{find_missing}
\alias{find_missing}
\title{Find missing simulations}
\usage{
find_missing(simulations, pattern = "^sim_", pb = TRUE)
}
\arguments{
\item{simulations}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{pb}{Whether to display a progress bar}
}
\description{
Find missing simulations
}
