% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing.R
\name{find_missing}
\alias{find_missing}
\title{Find missing simulations}
\usage{
find_missing(sims)
}
\arguments{
\item{sims}{A vector of simulation folders}
}
\value{
A character vector of missing simulations, or NULL if no simulation
is missing
}
\description{
Find missing simulations
}
\details{
Uses the absence of \code{.dat} output files as a clue for missing
simulations
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data"

roots <- fetch_dirs(root, "example_", level = 1)
find_missing(roots)

}

}
