% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing.R
\name{find_missing}
\alias{find_missing}
\title{Find missing simulations}
\usage{
find_missing(sims, pattern = "^sim_", verbose = TRUE, level = 0)
}
\arguments{
\item{sims}{Either path to a root directory containing simulation folders, or a vector of simulation folders}

\item{pattern}{Pattern defining the simulation folders to look into}

\item{verbose}{Whether to display messages}

\item{level}{Recursion level}
}
\description{
Find missing simulations
}
