% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degrees.R
\name{get_degrees}
\alias{get_degrees}
\title{Get locus degrees}
\usage{
get_degrees(arch)
}
\arguments{
\item{arch}{A genetic architecture}
}
\value{
A vector of degrees across loci
}
\description{
Get locus degrees
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

arch <- read_arch(root)
get_degrees(arch)

}

}
