% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{SLURM status}
\usage{
get_status(folder)
}
\arguments{
\item{folder}{Path to the folder}
}
\value{
A character
}
\description{
Read the exit status of the simulation from the SLURM output file
}
