% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_extinct.R
\name{is_extinct}
\alias{is_extinct}
\title{Is a simulation extinct?}
\usage{
is_extinct(root, logfile = "^log.txt$", slurm = FALSE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{logfile}{Optional log file name}

\item{slurm}{Whether to check a SLURM output file as log file}
}
\value{
A logical
}
\description{
Checks for the word "extinct" in the log file
}
\examples{

\dontrun{

is_extinct("data/example_1")

}

}
