% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_missing.R
\name{is_missing}
\alias{is_missing}
\title{Is a simulation missing?}
\usage{
is_missing(folder)
}
\arguments{
\item{folder}{Path to the folder}
}
\value{
A logical
}
\description{
Is a simulation missing?
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

is_missing(root)

}

}
