% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_facet_labels.R
\name{make_facet_labels}
\alias{make_facet_labels}
\title{Make facet labels}
\usage{
make_facet_labels(
  data,
  columns,
  prefixes = NULL,
  no_use = FALSE,
  sep = " = ",
  simplify = TRUE
)
}
\arguments{
\item{data}{The dataset}

\item{columns}{The variables to make facet labels for}

\item{prefixes}{Prefixe to add to each facet variable. Can be a named vector where the names are the corresponding facetting variables. Must be the same length as columns.}

\item{no_use}{Whether to return labels without prefix. Useful only in specific cases e.g. use inside other functions.}

\item{sep}{Separator between prefix and label. Defaults to " = ".}

\item{simplify}{If one facetting variable is supplied, whether to return a vector rather than a list of one vector}
}
\value{
A named list of named vectors of labels for each facetting variable. The output is ready for use as the "labeller" argument of the facet_wrap and facet_grid functions in ggplot by using labeller = do.call("labeller", labels), where labels is the output of this function. Returns a vector of named labels if only one facetting variable is requested and simplify is TRUE. In this case use labeller = labeller(ecosel = labels), assuming that labels is the function output and ecosel is the name of the facetting variable.
}
\description{
Make facet labels of the form "variable = value".
}
