% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrep.R
\name{mrep}
\alias{mrep}
\title{Repeat multiple elements multiple times}
\usage{
mrep(x, n)
}
\arguments{
\item{x}{Vector of what to repeat}

\item{n}{Vector of how many times}
}
\description{
Repeat multiple elements multiple times
}
