% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_heatmap.R
\name{plot_genome_heatmap}
\alias{plot_genome_heatmap}
\title{Genome heatmap}
\usage{
plot_genome_heatmap(folder, variable)
}
\arguments{
\item{folder}{The simulation folder}

\item{variable}{What variable to plot}
}
\value{
A ggplot
}
\description{
Plots a heatmap of a variable across the genome, through time
}
