% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_ridges.R
\name{plot_genome_ridges}
\alias{plot_genome_ridges}
\title{Genome-wide distributions through time}
\usage{
plot_genome_ridges(root, y, times = NULL, archfile = "architecture.txt")
}
\arguments{
\item{root}{Path to the simulation}

\item{y}{Variable to show}

\item{times}{Optional vector of time points to show}

\item{archfile}{Optional architecture file name}
}
\value{
A ggplot
}
\description{
Genome-wide distributions through time
}
