% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_scan.R
\name{plot_genome_scan}
\alias{plot_genome_scan}
\title{Genome scan plot}
\usage{
plot_genome_scan(
  data,
  mode = "timepoint",
  t = NULL,
  x = "location",
  y = "genome_Fst",
  show_generation = FALSE,
  style = "lollipop",
  color = NULL,
  show_chroms = FALSE,
  chrom_pars = list(y = -0.5, offset = -0.01, size = 2, color = "black")
)
}
\arguments{
\item{data}{A data frame containing information on a per locus (and optionally per time point) basis}

\item{mode}{Either of "timepoint" to filter a specific time point, "summary" if there are no multiple time points, and "animation" if an animation through time points should be made}

\item{t}{Integer. The time point to filter}

\item{x}{String. Either of "location" to show loci locations along the genome as the horizontal axis, or "locus" to show loci indices}

\item{y}{String. Name of the variable on the vertical axis}

\item{show_generation}{Logical. Whether to show the time point as a title}

\item{style}{String. The \code{geom}s to use. Either of "points", "bars" or "lollipop".}

\item{color}{Name of the variables to map to those aesthetics.}

\item{chrom_pars}{List of graphical parameters for the chromosomes, to be passed to \code{add_chromosomes()}. Includes \code{y}, \code{offset}, \code{size} and \code{color}.}
}
\value{
A ggplot, or an animation from the \code{gganimate} package
}
\description{
Flexible function to plot a variable across loci along the genome.
The data can be plotted for a given time point in a simulation,
or for no specific time point if the data are not timed (e.g. summary data).
The plot can be animated through time points. Edges of the gene network
can be displayed onto the layer of points representing the loci.
Supports additional aesthetics mapping and customization.
}
\note{
This is a simplified version of \code{plot_genome_scatter()}
}
