% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_violin.R
\name{plot_genome_violin}
\alias{plot_genome_violin}
\title{Diagnostic violin plot}
\usage{
plot_genome_violin(
  root,
  y,
  x = NULL,
  t = NULL,
  is_color = FALSE,
  archfile = "architecture.txt"
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Name of the variable to plot}

\item{x}{Optional grouping variable}

\item{t}{Time point at which to show the scan. If unspecified, last time point.}

\item{is_color}{Whether to color by \code{x}}

\item{archfile}{Optional architecture file name}
}
\value{
A ggplot
}
\description{
Quick-plots a variable across the genome, possibly split between categories
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_genome_violin(root, "genome_Fst")
plot_genome_violin(root, "genome_Fst", x = "trait", is_color = TRUE)

}
