% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_line.R
\name{plot_sim_line}
\alias{plot_sim_line}
\title{Plot a simulation-wise variable}
\usage{
plot_sim_line(root, y, x = "time", by = 1, j = 1, color = "black")
}
\arguments{
\item{root}{Path to the simulation folder}

\item{y}{Variable to track}

\item{x}{Optional variable for the horizontal axis. Defaults to "time".}

\item{j}{If the variable is splitted, which column to show?}

\item{color}{Optional color of the line}

\item{bins}{Number of bins for \code{geom_bin2d}}
}
\value{
A ggplot
}
\description{
Plot a simulation-wise variable
}
\examples{

root <- system.file("extdata", "example_1", package = "egssimtools")
plot_sim_line(root, "EI")

}
