% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_density.R, R/plot_trait_density2D.R
\name{plot_trait_density}
\alias{plot_trait_density}
\title{Heatmap of trait values through time}
\usage{
plot_trait_density(folder, traits = c(1, 2), t = NULL, bins = 15)

plot_trait_density(folder, traits = c(1, 2), t = NULL, bins = 15)
}
\arguments{
\item{traits}{Vector of two traits to plot: (1) ecological trait, (2) mating preference and (3) neutral trait.}

\item{t}{What time point to show. All time points together if NULL.}

\item{bins}{Number of bins in the heatmap}

\item{trait}{What trait to plot? (1) Ecological trait, (2) mating preference and (3) neutral trait.}

\item{filename}{Path to the folder}
}
\value{
A ggplot with the basic needed components. You can add extra layers to it (e.g. change the axis labels) outside the function.

A ggplot with the basic needed components. You can add extra layers to it (e.g. change the axis labels) outside the function.
}
\description{
Shows the densities of trait values through time.

Shows the joint densities of two traits across the population.
}
