% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_values.R
\name{plot_trait_values}
\alias{plot_trait_values}
\title{Plot trait values through time}
\usage{
plot_trait_values(data, bins = 30)
}
\arguments{
\item{data}{An individual-wise dataset where each trait is a different column.}

\item{bins}{Number of bins to pass to \code{ggplot2::geom_bin2d}. Defaults to 30.}
}
\value{
A ggplot
}
\description{
Density heatmap of the three trait values across the population.
}
