% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_archfile.R
\name{read_archfile}
\alias{read_archfile}
\title{Read an architecture file}
\usage{
read_archfile(filename)
}
\arguments{
\item{filename}{Path to the architecture file}
}
\value{
A list with all the fields of a genetic architecture
}
\description{
Read an architecture file
}
\examples{

f <- system.file("extdata", "example_1", package = "egssimtools")
f <- file.path(f, "architecture.txt")
read_archfile(f)

}
