% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population.R
\name{read_population}
\alias{read_population}
\title{Read population-wide data}
\usage{
read_population(folder, variable, nvalues = 1, add_time = FALSE)
}
\arguments{
\item{variable}{What to read}

\item{nvalues}{Number of values per time point}

\item{add_time}{Whether to try to append a time vector to the result}

\item{filename}{Path to the folder}
}
\description{
Read population-wide data
}
