% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sim.R
\name{read_sim}
\alias{read_sim}
\title{Read simulation summary through time}
\usage{
read_sim(
  folder,
  variables,
  by = rep(1, length(variables)),
  parnames = NULL,
  combine = FALSE,
  as_numeric = NULL,
  parfile = "paramlog.txt"
)
}
\arguments{
\item{folder}{Path to the folder}

\item{variables}{What variables to read (\code{time} is included by default)}

\item{by}{Same as the \code{by} argument of \code{read_data} (do not count \code{time} in)}

\item{parnames, combine, as_numeric, parfile}{Parameters for \code{read_data}}
}
\value{
A tibble
}
\description{
Wrapper around \code{read_data} to read timepoint-wise summary data
}
\details{
The file \code{time.dat} must be present
}
\examples{

\dontrun{

# Location of the simulation folder
root <- "data/example_1"

# Read the ecological divergence through time
read_sim(root, "EI")

# Read multiple speciation metrics
read_sim(root, c("EI", "RI"))

# Read genome-wide metrics
read_sim(root, "Fst", by = 3)

# Or a combination of multiple metrics with different dimensions
read_sim(root, c("EI", "Fst"), by = c(1, 3))

# Even locus-specific data in the wide-format
read_sim(root, "genome_Fst", by = 300)

}

}
